+++ Packs/RiskSense/Integrations/RiskSense/CHANGELOG.md
@@ -1,6 +1,15 @@
 ## [Unreleased]
--
-
+ - New Features
+    - 3 new playbooks
+        - Ransomware exposure
+        - BlueKeep CVE exposure
+        - DHS-CISA-FBI Top 10 CVE exposure
+    - Helper script for ransomware exposure
+    - risksense-apply-tag command
+ - Enhancement
+    - new RiskSense branding logo
+    - minor argument name change
+    - minor context data change
 
 ## [20.4.0] - 2020-04-14
 -